app/requirements.txt
amqp==5.0.3
apns2==0.7.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.5
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==2.9.2
Django==3.0.8
django-easymode==0.14.5
django-fcm==0.1.1
django-jsonfield==1.4.0
django-modeltranslation==0.16
django-multiselectfield==0.1.12
django-push-notifications==2.0.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
kombu==5.0.2
msgpack==0.6.1
Pillow==7.2.0
prompt-toolkit==3.0.10
psycopg2==2.8.5
py-vapid==1.7.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.1
pytz==2020.1
pywebpush==1.11.0
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
text-unidecode==1.3
timeago==1.0.14
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5